Music: Fix drag progress bar will hear noise in file manager
The seekbar of audiopreview will update the progress of player every
time that it detects the progress change of itself. Now the seek action
of player will take place only when onStopTrackingTouch is triggered.
The same problem also exists in Music app. So the change is also
applied to it.

Change-Id: Ie263c380f149c19309d44a354ae8c4a5e5c41a82
Signed-off-by: Xiaojing Zhang <zhangx@codeaurora.org>